stocks_server/requirements.txt
asttokens==2.4.0
backcall==0.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
debugpy==1.7.0
decorator==5.1.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.1.3
executing==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.3
jupyter_client==8.3.1
jupyter_core==5.3.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
more-itertools==8.10.0
nest-asyncio==1.5.7
netifaces==0.11.0
oauthlib==3.2.0
oracledb==2.1.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==5.4.1
pyzmq==25.1.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.28
stack-data==0.6.2
systemd-python==234
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==3.0.1
zipp==1.0.0
